Improvements and adaptations to work with HomeDC
In the case where a HomeDC-like deployment is made
behind a SOCKS proxy with limited network interfaces,
containerization of chameleonsocks only creates more
headaches.

This commit makes it easier to run chameleonsocks
"locally", as in not in a container, and also
in secondary network namespaces. For the other piece
of the puzzle, check chameleonsocks-kolla-friendly.sh
in the homedc repository.

Also some improvements and more proxy exceptions by default.

Original chameleonsocks script file by Todor Minchev
has been renamed to chameleonsocks-docker.

And make it run without Docker as my use case
requires chameleonsocks to be run on the same
netspace as one of OpenStack's L3 router.